+++ Nickel/Models/Content/DeckConfiguration.cs
@@ -22,8 +22,24 @@ public readonly struct DeckConfiguration
 	/// <summary>A localization provider for the name of the <see cref="Deck"/>.</summary>
 	public SingleLocalizationProvider? Name { get; init; }
 	
+	/// <summary>A delegate that can override the frame of the card on a per-card basis.</summary>
+	public Func<CardFrameOverrideArgs, Spr>? CardFrameOverride { get; init; }
+	
 	/// <summary>A delegate that can override the default rarity shine of a card.</summary>
 	public Func<ShineColorOverrideArgs, Color>? ShineColorOverride { get; init; }
+
+	/// <seealso cref="CardFrameOverride"/>
+	public struct CardFrameOverrideArgs
+	{
+		/// <summary>The current state of the game.</summary>
+		public required State State { get; init; }
+		
+		/// <summary>The card being rendered.</summary>
+		public required Card Card { get; init; }
+		
+		/// <summary>The default card frame sprite for this card.</summary>
+		public required Spr DefaultFrameSprite { get; init; }
+	}
 
 	/// <seealso cref="ShineColorOverride"/>
 	public struct ShineColorOverrideArgs
@@ -43,6 +59,9 @@ public readonly struct DeckConfiguration
 	/// </summary>
 	public struct Amends
 	{
+		/// <inheritdoc cref="DeckConfiguration.ShineColorOverride" />
+		public ContentConfigurationValueAmend<Func<CardFrameOverrideArgs, Spr>?>? CardFrameOverride { get; set; }
+		
 		/// <inheritdoc cref="DeckConfiguration.ShineColorOverride" />
 		public ContentConfigurationValueAmend<Func<ShineColorOverrideArgs, Color>?>? ShineColorOverride { get; set; }
 	}